+++ config_manager/config_defaults.h
@@ -72,6 +72,8 @@ namespace cfg {
 
 	extern unsigned leds_brightness;
 	extern bool leds_on;
+	extern unsigned leds_off_hour;
+	extern unsigned leds_on_hour;
 
 	// data sharing preferences
 	extern bool share_temperature;