+++ display/paint_driver/fonts/fonts.h
@@ -95,11 +95,13 @@ extern const Font font_12_cyrillic;
 extern const Font font_16_cyrillic;
 extern const Font font_20_cyrillic;
 extern const Font font_24_cyrillic;
+extern const Font font_32_cyrillic;
 extern const Font font_8_ascii;
 extern const Font font_12_ascii;
 extern const Font font_16_ascii;
 extern const Font font_20_ascii;
 extern const Font font_24_ascii;
+extern const Font font_32_ascii;
 
 extern cFONT Font12CN;
 extern cFONT Font24CN;